reat: CoordinatedReader shared fetch (#44)
* refactor(core,sql): bump partition id/count from u16 to u32

* feat(sqlite): add SqlitePartitionBackfill for partition backfill parity with Postgres

* refactor(core/io/reader): concurrent partition release on CoordinatedReader shutdown

* refactor(core/io): with_partitions becomes primary PartitionableSubscription method

* refactor(postgres,sqlite): carry Partition in PgCursor and SqliteCursor

* feat(core/io/reader): HasPartition trait and PartitionedReader::source_from_cursor

* feat(core/io/reader): batched checkpoint flush in CoordinatedReader

* refactor(core/io/reader): shared-fetch data plane in CoordinatedReader

* docs: parallelism and multi-instance after shared-fetch landing

* test(postgres): tolerate OwnershipLost in concurrent-owners test

+++ crates/eventuary-postgres/src/event_log.rs
@@ -23,8 +23,8 @@ CREATE TABLE IF NOT EXISTS {events} (
     causation_id TEXT,
     partition_key TEXT,
     partition_hash BIGINT,
-    partition_id INT,
-    partition_count INT,
+    partition_id BIGINT,
+    partition_count BIGINT,
     partition_strategy TEXT
 );
